+++ src/main/java/io/github/kirstenali/deepj/loss/LossFunction.java
@@ -3,6 +3,6 @@ package io.github.kirstenali.deepj.loss;
 import io.github.kirstenali.deepj.tensor.Tensor;
 
 public interface LossFunction {
-    double loss(Tensor predicted, Tensor actual);
+    float loss(Tensor predicted, Tensor actual);
     Tensor gradient(Tensor predicted, Tensor actual);
 }
