+++ src/main/java/io/github/kirstenali/deepj/loss/MSELoss.java
@@ -4,17 +4,17 @@ import io.github.kirstenali.deepj.tensor.Tensor;
 
 public class MSELoss implements LossFunction {
     @Override
-    public double loss(Tensor predicted, Tensor actual) {
+    public float loss(Tensor predicted, Tensor actual) {
         Tensor diff = predicted.subtract(actual);
         Tensor sq = diff.multiply(diff);
-        double sse = sq.sum();
+        float sse = sq.sum();
         return sse / (predicted.rows * predicted.cols);
     }
 
     @Override
     public Tensor gradient(Tensor predicted, Tensor actual) {
         Tensor grad = predicted.subtract(actual);
-        grad.multiplyScalarInPlace(2.0 / (predicted.rows * predicted.cols));
+        grad.multiplyScalarInPlace(2.0f / (predicted.rows * predicted.cols));
         return grad;
     }
 }
