+++ src/main/java/io/github/kirstenali/deepj/models/gpt/GPTConfig.java
@@ -9,17 +9,17 @@ public record GPTConfig(
         int nHeads,
         int nLayers,
         int dFF,
-        double initScale,
-        double gradClipNorm
+        float initScale,
+        float gradClipNorm
 ) implements TransformerConfig {
 
     public GPTConfig(int vocabSize, int maxSeqLen, int dModel, int nHeads, int nLayers, int dFF) {
-        this(vocabSize, maxSeqLen, dModel, nHeads, nLayers, dFF, 0.2, 1.0);
+        this(vocabSize, maxSeqLen, dModel, nHeads, nLayers, dFF, 0.2f, 1.0f);
     }
 
     public GPTConfig {
         TransformerConfig.validateCommon(vocabSize, maxSeqLen, dModel, nHeads, nLayers, dFF, gradClipNorm);
-        if (!Double.isFinite(initScale) || initScale <= 0.0)
+        if (!Float.isFinite(initScale) || initScale <= 0.0f)
             throw new IllegalArgumentException("initScale must be finite and > 0");
     }
 }
