+++ src/main/java/io/github/kirstenali/deepj/training/TrainingResult.java
@@ -1,3 +1,3 @@
 package io.github.kirstenali.deepj.training;
 
-public record TrainingResult(int steps, double lastLoss, double emaLoss) {}
+public record TrainingResult(int steps, float lastLoss, float emaLoss) {}
